Reviewed walletService
