+++ app/src/main/java/com/tari/android/wallet/service/connection/TariWalletServiceConnection.kt
@@ -40,7 +40,7 @@ import android.os.IBinder
 import androidx.lifecycle.ViewModel
 import com.tari.android.wallet.application.TariWalletApplication
 import com.tari.android.wallet.service.TariWalletService
-import com.tari.android.wallet.service.WalletService
+import com.tari.android.wallet.service.service.WalletService
 import io.reactivex.Observable
 import io.reactivex.subjects.BehaviorSubject
