+++ app/src/main/java/com/tari/android/wallet/service/service/BaseNodeValidationType.kt
@@ -1,0 +1,6 @@
+package com.tari.android.wallet.service.service
+
+enum class BaseNodeValidationType {
+    TXO,
+    TX;
+}
