+++ app/src/main/java/com/tari/android/wallet/ui/fragment/auth/AuthViewModel.kt
@@ -3,7 +3,7 @@ package com.tari.android.wallet.ui.fragment.auth
 import com.tari.android.wallet.data.sharedPrefs.SharedPrefsRepository
 import com.tari.android.wallet.data.sharedPrefs.network.NetworkRepository
 import com.tari.android.wallet.infrastructure.security.biometric.BiometricAuthenticationService
-import com.tari.android.wallet.service.WalletServiceLauncher
+import com.tari.android.wallet.service.service.WalletServiceLauncher
 import com.tari.android.wallet.ui.common.CommonViewModel
 import javax.inject.Inject
