+++ app/src/main/java/com/tari/android/wallet/ui/fragment/settings/deleteWallet/DeleteWalletActivity.kt
@@ -40,7 +40,7 @@ import androidx.lifecycle.lifecycleScope
 import com.tari.android.wallet.R
 import com.tari.android.wallet.databinding.ActivityDeleteWalletBinding
 import com.tari.android.wallet.di.DiContainer.appComponent
-import com.tari.android.wallet.service.WalletServiceLauncher
+import com.tari.android.wallet.service.service.WalletServiceLauncher
 import com.tari.android.wallet.service.connection.TariWalletServiceConnection
 import com.tari.android.wallet.ui.common.CommonActivity
 import com.tari.android.wallet.ui.dialog.modular.DialogArgs
